requirements.txt
jinja2
pyyaml
typer